Initial port to Gtk3 SL #3681
This is a big change that ports Get Books to Gtk3. Here are the steps[1]
that I followed to this port.

Pending issues (some will be addressed in the theme):

 * Cancel search button. It shows a really big X when you start
typing: http://bugs.sugarlabs.org/ticket/3385

 * Clicking some things on the screen makes the ListView grows a bit
and it doesn't fit on the screen (the right scroll-bar goes outside
the window)

 * Some time when I click on the Search Entry all the toolbar is moved a bit

 * When the "generic book image" is shown (at the bottom
left), the gray color used to fill it is not the same as the
background. I changed the color with is the GdkPixbuf filled to match
with the darker one (style.COLOR_BUTTON_GREY) because I couldn't find
how to fill the alpha of the "generic_cover.png" with the old color
(style.COLOR_PANEL_GREY)

 * After searching and selecting a book, if we change the format to
EPUB the Get Book button is moved a bit

 * When the book download starts, a Cancel is shown but it doesn't
have the X icon

 * The background of the rows on the ListView do not cycle between
white and gray

 * The Catalogs column title is bigger than the other column titles

 * The column width should be re-sized to fit with the text inside
them with a max width pre-set

 * Right clicking on the column titles doesn't show the Gtk.Menu popup

 * UDisk has changed the interface, then is not working

 * A error with the display of book description using the activity in spanish
due to mismanagement of special chars.

[1] http://wiki.sugarlabs.org/go/User:Humitos/PortingGetBooks

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Reviewed-by: Gonzalo Odiard <gonzalo@laptop.org>