Fixing the widening of the download box canvas.
The icons at the bottom left corner of the screen used to get hidden, moving
towards right into the screen and so does the stop button, whenever a book
with long Summary is called.

By decreasing the left and right margin of the textview of the scrol window
from "20" to "5" this issue is fixed.

Fixes https://github.com/sugarlabs/get-books-activity/issues/24