Fixing no internet connection exception
When the activity is not connected to internet no alert message was shown,
although there is an exception handling function implemented in the code,
named "__query_updated_cb" but it had some errors.

"bozo_exception.reason" need not to be an instance of gai(get address info) error
always so changing socket.gaierror to socket.error and
bozo_exception.reason.errno to 101.

Fixes https://github.com/sugarlabs/get-books-activity/issues/28

sa