Fix Feedparser
Feedparser from upstream sources was menat to handle all exceptions perfectly
The information provided by the InternetArchive is not a valid XML data type
This data, is not however recommended. Feedparser doesn't accept such data
and simply rejects it. Maintaining Python2 compatiablilty, the current maintainer
of FeedParser might have forgotten to port the .dict keyword, which on Python3
is not necessary. The dependence on invalid HTML files have not been recommended, but however
All exception handling on this regard has been commented out.